// backend/migrations/snapshots/b024885b9498e1596c5d1c2c70179e7c887fe4d25fe33005237a212c94a4e62e/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Home: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly timezone: CodecTypes['pg/text@1']['output'];
      readonly archivedAt: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Membership: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly homeId: CodecTypes['pg/uuid@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
      readonly role: 'ROOMMATE' | 'ADMIN';
      readonly joinedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly endedAt: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
    };
    readonly User: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Home: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly timezone: CodecTypes['pg/text@1']['input'];
      readonly archivedAt: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Membership: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly homeId: CodecTypes['pg/uuid@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
      readonly role: 'ROOMMATE' | 'ADMIN';
      readonly joinedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly endedAt: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
    };
    readonly User: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly homes: {
      readonly archived_at: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly timezone: CodecTypes['pg/text@1']['output'];
      readonly updated_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly memberships: {
      readonly ended_at: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly home_id: CodecTypes['pg/uuid@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly joined_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly role: 'ROOMMATE' | 'ADMIN';
      readonly user_id: CodecTypes['pg/uuid@1']['output'];
    };
    readonly users: {
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly updated_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly homes: {
      readonly archived_at: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly timezone: CodecTypes['pg/text@1']['input'];
      readonly updated_at: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly memberships: {
      readonly ended_at: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly home_id: CodecTypes['pg/uuid@1']['input'];
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly joined_at: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly role: 'ROOMMATE' | 'ADMIN';
      readonly user_id: CodecTypes['pg/uuid@1']['input'];
    };
    readonly users: {
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly updated_at: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;